+++ packages/row_layout/CHANGELOG.md
@@ -1,6 +1,9 @@
+## 1.0.2
+* Made the row and columns to display empty message when blank
+* Made the main widget as Stack instead of wrapping with Container widget
+
 ## 1.0.1
-
 * Documentation, website and git urls updated
+* 
 ## 1.0.0
-
 * Simple and powerful flutter layout